extract `syntax.rs` to `rep_lang_core`
