+++ rep_lang_parser/Cargo.toml
@@ -11,6 +11,7 @@ combine = "4.3.2"
 pretty = "0.10.0"
 quickcheck = "0.9.2"
 rand = "0.7"
+rep_lang_core = { path = "../rep_lang_core" }
 
 [dev-dependencies]
 quickcheck_macros = "0.9.1"
