+++ rep_lang_parser/src/parse.rs
@@ -3,7 +3,7 @@ use combine::parser::char::{alpha_num, char, digit, letter, spaces, string};
 use combine::stream::{Stream, StreamErrorFor};
 use combine::{attempt, between, choice, many, many1, not_followed_by, optional, parser, Parser};
 
-use super::syntax::*;
+use rep_lang_core::syntax::*;
 
 // `impl Parser` can be used to create reusable parsers with zero overhead
 pub fn expr_<Input>() -> impl Parser<Input, Output = Expr>
