+++ rep_lang_runtime/Cargo.toml
@@ -11,6 +11,7 @@ license = "MIT"
 combine = "4.3.2"
 pretty = "0.10.0"
 rustyline = "6.3.0"
+rep_lang_core = { path = "../rep_lang_core" }
 rep_lang_parser = { path = "../rep_lang_parser" }
 
 [dev-dependencies]
